+++ src/main.rs
@@ -1,7 +1,7 @@
 use std::convert::AsRef;
 use std::fmt::Debug;
 
-mod library_mod;
+mod library;
 mod race;
 mod utils;
 
@@ -248,9 +248,9 @@ fn life_time_2() {
 }
 
 fn library() {
-  let mut library = library_mod::library::Library::new();
+  let mut library = library::entity::Library::new();
 
-  let favorite_book = library_mod::library::Book::new("A sample book", 1985);
+  let favorite_book = library::entity::Book::new("A sample book", 1985);
   println!(
     "Our favorite book {} should go in our library.",
     favorite_book
@@ -260,7 +260,7 @@ fn library() {
   println!("3: {:p}", &library); // same pointer
   println!("Our library: {library:?}.");
 
-  let favorite_book_2 = library_mod::library::Book::new("A sample book 2", 1986);
+  let favorite_book_2 = library::entity::Book::new("A sample book 2", 1986);
   library.add_book(favorite_book_2); // same pointer
   println!("Our library 2: {library:?}.");
